+++ src/steps/resource-manager/constants.ts
@@ -1,0 +1,2 @@
+export const ANY_SCOPE = 'ANY_SCOPE';
+export const ANY_RESOURCE = 'ANY_RESOURCE';